Refactor and add documentation
